Refactor service

+++ src/main/kotlin/tech/jaya/ridely/util/RidelyUtil.kt
@@ -5,7 +5,7 @@ import java.math.RoundingMode
 
 class RideUtil {
 
-    fun calculaPrice(kmDistance: Double, timeMin: Int): BigDecimal {
+    fun calculaPrice(kmDistance: Double, timeMin: Double): BigDecimal {
         val price = (kmDistance * 3.0 + timeMin * 2.0).toBigDecimal().setScale(2, RoundingMode.HALF_EVEN)
         val taxaApp = price.multiply(BigDecimal("0.01")).setScale(2, RoundingMode.HALF_EVEN)
